ENH: lh#dev#class#ancestors() overloaded
The function accepts a new overload that takes a list of base classes in
order to find all base classes.
The typical scenario is code generation: we are defining a new class, we
know its parents, but as the class hasn't been generated/written yet,
the tags DB has no reference on the class being constructed.

See lh-cpp cpp/clonable-class snippet/wizard